Correct merge crash from long ago.
Simplepie.inc.orig had merge errors form a commit long ago.  Didn't seem
to affect performance (probably because the functional file is actually
simplepie.inc), but the "orig" was throwing an error in Netbeans due to
all the >>>>> and <<<<< lying about.  There didn't seem to be any real
damage, just a bunch of undeleted HEAD and commit hash tags.